explorer: Synchronoize timescale for tps and ping stats
