+++ explorer/src/pages/ClusterStatsPage.tsx
@@ -9,7 +9,7 @@ import {
 } from "providers/stats/solanaClusterStats";
 import { abbreviatedNumber, lamportsToSol, slotsToHumanString } from "utils";
 import { ClusterStatus, useCluster } from "providers/cluster";
-import { TpsCard } from "components/TpsCard";
+import { LiveTransactionStatsCard } from "components/LiveTransactionStatsCard";
 import { displayTimestampWithoutDate } from "utils/date";
 import { Status, useFetchSupply, useSupply } from "providers/supply";
 import { ErrorCard } from "components/common/ErrorCard";
@@ -18,7 +18,6 @@ import { useVoteAccounts } from "providers/accounts/vote-accounts";
 import { CoingeckoStatus, useCoinGecko } from "utils/coingecko";
 import { Epoch } from "components/common/Epoch";
 import { TimestampToggle } from "components/common/TimestampToggle";
-import { SolanaPingCard } from "components/SolanaPingCard";
 
 const CLUSTER_STATS_TIMEOUT = 5000;
 
@@ -36,8 +35,7 @@ export function ClusterStatsPage() {
         </div>
         <StatsCardBody />
       </div>
-      <TpsCard />
-      <SolanaPingCard />
+      <LiveTransactionStatsCard />
     </div>
   );
 }
